relay-controller/requirements.txt
Flask
flask-cors
RPi.GPIO
simple_pid
adafruit-circuitpython-sht31d
redis
paho-mqtt
git+https://github.com/benoitc/gunicorn.git